refactor: flatten crypto module structure
- Move algebra module from primitives to top level
- Remove primitives module entirely
- Relocate threshold and vss modules
- Update all import paths throughout codebase
