+++ src/crypto.rs
@@ -1,5 +1,7 @@
+pub mod algebra;
 pub mod dkg;
 pub mod index_map;
 pub mod keypair;
-pub mod primitives;
+pub mod threshold;
 pub mod tss;
+pub mod vss;
