+++ src/crypto/dkg.rs
@@ -1,16 +1,29 @@
+use std::collections::HashSet;
+
 use crate::{
     crypto::{
+        algebra::{Point, Scalar},
         index_map::IndexedMap,
         keypair::PublicKey,
-        primitives::algebra::{Point, Scalar},
     },
     mocks::MockError,
 };
-use std::{collections::HashSet, error::Error};
-
-use mockall::automock;
 
 pub mod joint_feldman;
+
+pub use joint_feldman::JointFeldman;
+
+#[mockall::automock(type Error=MockError;)]
+#[async_trait::async_trait]
+pub trait Dkg {
+    type Error: std::error::Error;
+
+    async fn set_peers(
+        &mut self,
+        peers: IndexedMap<libp2p::PeerId, PublicKey>,
+    ) -> Result<(), Self::Error>;
+    async fn generate(&self, id: Vec<u8>) -> Result<GenerateResult, Self::Error>;
+}
 
 pub enum GenerateResult {
     Success {
@@ -20,16 +33,4 @@ pub enum GenerateResult {
     Failure {
         invalid_peers: HashSet<libp2p::PeerId>,
     },
-}
-
-#[automock(type Error=MockError;)]
-#[async_trait::async_trait]
-pub trait Dkg {
-    type Error: Error;
-
-    async fn set_peers(
-        &mut self,
-        peers: IndexedMap<libp2p::PeerId, PublicKey>,
-    ) -> Result<(), Self::Error>;
-    async fn generate(&self, id: Vec<u8>) -> Result<GenerateResult, Self::Error>;
 }
