--- src/crypto/primitives.rs
@@ -1,3 +1,0 @@
-pub mod algebra;
-pub mod threshold;
-pub mod vss;
