+++ src/crypto/threshold/counter.rs
