+++ src/network/transport/libp2p_transport/protocols/gossipsub/payload.rs
@@ -3,12 +3,10 @@ use std::collections::HashSet;
 use serde::{Deserialize, Serialize};
 
 use crate::crypto::{
+    algebra::{Point, Scalar},
     index_map::IndexedMap,
     keypair::{PublicKey, VrfProof},
-    primitives::{
-        algebra::{Point, Scalar},
-        vss::{encrypted_share::EncryptedShares, DecryptedShares},
-    },
+    vss::{encrypted_share::EncryptedShares, DecryptedShares},
 };
 
 #[derive(Debug)]
